[on-chain-config] on chain config by dao package block
