+++ executor/src/stdlib_test/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-pub use module_upgrade_test::dao_vote_test;
-
+#[cfg(test)]
 mod module_upgrade_test;
+#[cfg(test)]
 mod on_chain_config_test;
