+++ test-helper/src/lib.rs
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod chain;
+pub mod dao;
 pub mod dummy_network_service;
+pub mod executor;
 pub mod network;
 pub mod node;
 pub mod txn;
@@ -12,6 +14,7 @@ pub use chain::gen_blockchain_for_test;
 pub use dummy_network_service::DummyNetworkService;
 pub use network::build_network;
 pub use node::{run_node_by_config, run_test_node};
+pub use starcoin_executor::Account;
 pub use starcoin_genesis::{Genesis, GenesisOpt};
 pub use starcoin_node::NodeHandle;
 pub use txpool::{start_txpool, start_txpool_with_size};
